Changed all commands to dedicated folders and better structure. Also added tests for the new search functionality.
